+++ pkg/types/artifact.go
@@ -24,7 +24,7 @@ package types
 
 // ArtifactItem represents an artifact.
 type ArtifactItem struct {
-	ID        uint     `json:"id"`
+	ID        uint64     `json:"id"`
 	Digest    string   `json:"digest"`
 	Size      int64    `json:"size"`
 	Tags      []string `json:"tags"`
@@ -43,7 +43,7 @@ type ListArtifactRequest struct {
 
 // GetArtifactRequest represents the request to get an artifact.
 type GetArtifactRequest struct {
-	ID         uint   `json:"id" param:"id" validate:"required,number"`
+	ID         uint64   `json:"id" param:"id" validate:"required,number"`
 	Namespace  string `json:"namespace" param:"namespace" validate:"required,min=2,max=20,is_valid_namespace"`
 	Repository string `json:"repository" query:"repository" validate:"required,is_valid_repository"`
 	Digest     string `json:"digest" param:"digest" validate:"required,is_valid_digest"`
@@ -51,7 +51,7 @@ type GetArtifactRequest struct {
 
 // DeleteArtifactRequest represents the request to delete an artifact.
 type DeleteArtifactRequest struct {
-	ID         uint   `json:"id" param:"id" validate:"required,number"`
+	ID         uint64   `json:"id" param:"id" validate:"required,number"`
 	Namespace  string `json:"namespace" param:"namespace" validate:"required,min=2,max=20,is_valid_namespace"`
 	Repository string `json:"repository" query:"repository" validate:"required,is_valid_repository"`
 	Digest     string `json:"digest" param:"digest" validate:"required,is_valid_digest"`
